Solving timeout problem and refactored tests
Fixes #162
