+++ selenium_tests/test_start_stop_container.py
@@ -1,32 +1,11 @@
 # -*- coding: utf-8 -*-
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.support.ui import Select
-from selenium.common.exceptions import NoSuchElementException
-from selenium.common.exceptions import NoAlertPresentException
 import unittest
 import time
-import re
+
+from selenium_tests.selenium_test_base import SeleniumTestBase
 
 
-class TestStartStopContainer(unittest.TestCase):
-
-    def setUp(self):
-        ff_binary = webdriver.firefox.firefox_binary.FirefoxBinary()
-        ff_profile = webdriver.firefox.firefox_profile.FirefoxProfile()
-        ff_profile.assume_untrusted_cert_issuer = True
-        ff_profile.accept_untrusted_certs = True
-        capabilities = webdriver.DesiredCapabilities().FIREFOX
-        capabilities['acceptSslCerts'] = True
-        self.driver = webdriver.Firefox(firefox_binary=ff_binary,
-                                        firefox_profile=ff_profile,
-                                        capabilities=capabilities)
-        self.driver.implicitly_wait(30)
-        self.base_url = "https://127.0.0.1:8000/"
-        self.verificationErrors = []
-        self.accept_next_alert = True
-
+class TestStartStopContainer(SeleniumTestBase):
     def test_start_stop_container(self):
         driver = self.driver
         driver.get(self.base_url + "/hub/login")
@@ -47,57 +26,9 @@ class TestStartStopContainer(unittest.TestCase):
             self.fail("time out")
         driver.find_element_by_link_text("Close").click()
         driver.find_element_by_name("action").click()
-        for i in range(60):
-            try:
-                if "noVNC" == driver.title:
-                    break
-            except:
-                pass
-            time.sleep(1)
-        else:
-            self.fail("time out")
+        self.wait_for(lambda: "noVNC" == driver.title)
         driver.find_element_by_xpath("//i").click()
         driver.find_element_by_xpath("(//button[@name='action'])[2]").click()
-        for i in range(60):
-            try:
-                if "Start" == driver.find_element_by_name("action").text:
-                    break
-            except:
-                pass
-            time.sleep(1)
-        else:
-            self.fail("time out")
+        self.wait_for(
+            lambda: "Start" == driver.find_element_by_name("action").text)
         driver.find_element_by_id("logout").click()
-
-    def is_element_present(self, how, what):
-        try:
-            self.driver.find_element(by=how, value=what)
-        except NoSuchElementException as e:
-            return False
-        return True
-
-    def is_alert_present(self):
-        try:
-            self.driver.switch_to_alert()
-        except NoAlertPresentException as e:
-            return False
-        return True
-
-    def close_alert_and_get_its_text(self):
-        try:
-            alert = self.driver.switch_to_alert()
-            alert_text = alert.text
-            if self.accept_next_alert:
-                alert.accept()
-            else:
-                alert.dismiss()
-            return alert_text
-        finally:
-            self.accept_next_alert = True
-
-    def tearDown(self):
-        self.driver.quit()
-        self.assertEqual([], self.verificationErrors)
-
-if __name__ == "__main__":
-    unittest.main()
